% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mainDataOutput.R
\name{mod_mainData}
\alias{mod_mainData}
\title{mod_mainData server function}
\usage{
mod_mainData(input, output, session, data_reactives, fesdb, lang)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data_reactives}{reactives from modules}

\item{fesdb}{object to access the fes db}

\item{lang}{lang selected}
}
\description{
mod_mainData server function
}
