% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_vizInput.R
\name{mod_viz}
\alias{mod_viz}
\title{mod_viz server function}
\usage{
mod_viz(input, output, session, data_reactives, var_thes, lang, cache)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data_reactives}{reactives needed}

\item{var_thes}{thesauruses}

\item{lang}{lang value}

\item{cache}{cache_mem object to store the selected col}
}
\description{
mod_viz server function
}
