% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{translate_var}
\alias{translate_var}
\title{translate variable function}
\usage{
translate_var(id, version, scale, lang, variables_thesaurus)
}
\description{
translate the variable based on the lang selected
}
